remove excessing logging

+++ api/cbbiinfo_api.py
@@ -15,6 +15,7 @@ def cbbi_fetch(key: str) -> pd.DataFrame:
             'Value',
         ],
     )
-    df['Date'] = pd.to_datetime(df['Date'], unit='s').dt.tz_localize(None)
+    # Convert string timestamps to int before datetime conversion to avoid FutureWarning
+    df['Date'] = pd.to_datetime(df['Date'].astype(int), unit='s').dt.tz_localize(None)
 
     return df
